requirements.txt
# Apple Notes Telegram Bot - Requirements

# Telegram bot
python-telegram-bot>=20.0

# HTTP requests (for Ollama API)
requests

# Audio processing
numpy
soundfile
ffmpeg  # System dependency: brew install ffmpeg

# Voice transcription (one of these)
parakeet-mlx  # Apple Silicon (M1/M2/M3/M4)
faster-whisper  # Intel/any system (fallback)

# Configuration
PyYAML

# macOS menu bar
rumps

# Optional
send2trash
